config test files moved to lib
