+++ lib/config/config/__fixtures__/file.js
