+++ lib/config/npm/__fixtures__/01.json
