+++ lib/config/npm/__fixtures__/02.json
