+++ lib/config/npm/__fixtures__/next.json
