+++ lib/config/npm/__fixtures__/typescript.json
