+++ lib/config/npm/__fixtures__/vue-test-utils.json
