+++ test/workers/repository/process/lookup/index.spec.ts
@@ -1,12 +1,12 @@
 import nock from 'nock';
 import * as lookup from '../../../../../lib/workers/repository/process/lookup';
-import qJson from '../../../../config/npm/__fixtures__/01.json';
-import helmetJson from '../../../../config/npm/__fixtures__/02.json';
-import coffeelintJson from '../../../../config/npm/__fixtures__/coffeelint.json';
-import webpackJson from '../../../../config/npm/__fixtures__/webpack.json';
-import nextJson from '../../../../config/npm/__fixtures__/next.json';
-import vueJson from '../../../../config/npm/__fixtures__/vue.json';
-import typescriptJson from '../../../../config/npm/__fixtures__/typescript.json';
+import qJson from '../../../../../lib/config/npm/__fixtures__/01.json';
+import helmetJson from '../../../../../lib/config/npm/__fixtures__/02.json';
+import coffeelintJson from '../../../../../lib/config/npm/__fixtures__/coffeelint.json';
+import webpackJson from '../../../../../lib/config/npm/__fixtures__/webpack.json';
+import nextJson from '../../../../../lib/config/npm/__fixtures__/next.json';
+import vueJson from '../../../../../lib/config/npm/__fixtures__/vue.json';
+import typescriptJson from '../../../../../lib/config/npm/__fixtures__/typescript.json';
 import { mocked, getConfig } from '../../../../util';
 import { CONFIG_VALIDATION } from '../../../../../lib/constants/error-messages';
 import * as dockerVersioning from '../../../../../lib/versioning/docker';
